+++ src/main/java/cn/qiuye/gtmoremachine/data/lang/CreativeLang.kt
@@ -1,10 +1,17 @@
 package cn.qiuye.gtmoremachine.data.lang
 
+import cn.qiuye.gtmoremachine.data.lang.LangHandler.addCN
 import cn.qiuye.gtmoremachine.data.lang.LangHandler.addCNEN
 
 object CreativeLang {
 
     fun init() {
+        addCN("block.gtmoremachine.creative_energy_hatch", "§r创造模式能源仓")
+        addCN("block.gtmoremachine.creative_fluid_input_hatch", "创造模式输入仓")
+        addCN("block.gtmoremachine.creative_item_input_bus", "创造模式输入总线")
+        addCN("block.gtmoremachine.creative_laser_hatch", "§r创造模式激光靶仓")
+        addCN("item.gtmoremachine.creative_energy_cover", "创造能源覆盖板")
+        addCN("item.gtmoremachine.creative_fluid_cell", "创造流体单元")
         addCNEN("item.gtmoremachine.creative_fluid_cell.tooltip.1", "§2内部流体：§f%1\$s", "§2Fluid Stored: §f%s")
         addCNEN(
             "item.gtmoremachine.creative_fluid_cell.tooltip.2",
