+++ src/main/java/cn/qiuye/gtmoremachine/data/lang/CreativeModeTabsLang.kt
@@ -1,0 +1,12 @@
+package cn.qiuye.gtmoremachine.data.lang
+
+import cn.qiuye.gtmoremachine.data.lang.LangHandler.addCN
+
+object CreativeModeTabsLang {
+
+    fun init() {
+        addCN("itemGroup.gtmoremachine.creative", "创造物品")
+        addCN("itemGroup.gtmoremachine.more_machines", "更多机器")
+        addCN("itemGroup.gtmoremachine.wireless", "格雷无线科技")
+    }
+}
