--- src/main/java/cn/qiuye/gtmoremachine/data/lang/CuriosLang.kt
@@ -1,7 +1,0 @@
-package cn.qiuye.gtmoremachine.data.lang
-
-object CuriosLang {
-    fun init() {
-//        provider.add("curios.identifier.gtmoremachine_wet", "Energy Terminal")
-    }
-}
