--- src/main/java/cn/qiuye/gtmoremachine/data/lang/Tobefixed.kt
@@ -1,504 +1,0 @@
-package cn.qiuye.gtmoremachine.data.lang
-
-import cn.qiuye.gtmoremachine.data.lang.LangHandler.addCN
-
-object Tobefixed {
-
-    fun init() {
-        addCN("block.gtmoremachine.creative_energy_hatch", "§r创造模式能源仓")
-        addCN("block.gtmoremachine.creative_fluid_input_hatch", "创造模式输入仓")
-        addCN("block.gtmoremachine.creative_item_input_bus", "创造模式输入总线")
-        addCN("block.gtmoremachine.creative_laser_hatch", "§r创造模式激光靶仓")
-        addCN("block.gtmoremachine.empty_tier_capacity_component", "空无线电网容量组件")
-        addCN("block.gtmoremachine.ev_1024a_wireless_laser_source_hatch", "1024安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_1024a_wireless_laser_target_hatch", "1024安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_1048576a_wireless_laser_source_hatch", "1,048,576安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_1048576a_wireless_laser_target_hatch", "1,048,576安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_16384a_wireless_laser_source_hatch", "16384安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_16384a_wireless_laser_target_hatch", "16384安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_16777216a_wireless_laser_source_hatch", "16,777,216安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_16777216a_wireless_laser_target_hatch", "16,777,216安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_16a_wireless_energy_input_hatch", "16安§5EV§r无线能源仓")
-        addCN("block.gtmoremachine.ev_16a_wireless_energy_output_hatch", "16安§5EV§r无线动力仓")
-        addCN("block.gtmoremachine.ev_256a_wireless_laser_source_hatch", "256安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_256a_wireless_laser_target_hatch", "256安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_262144a_wireless_laser_source_hatch", "262,144安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_262144a_wireless_laser_target_hatch", "262,144安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_2a_wireless_energy_input_hatch", "§5EV§r无线能源仓")
-        addCN("block.gtmoremachine.ev_2a_wireless_energy_output_hatch", "§5EV§r无线动力仓")
-        addCN("block.gtmoremachine.ev_33554432a_wireless_laser_source_hatch", "33,554,432安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_33554432a_wireless_laser_target_hatch", "33,554,432安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_4096a_wireless_laser_source_hatch", "4096安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_4096a_wireless_laser_target_hatch", "4096安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_4194304a_wireless_laser_source_hatch", "4,194,304安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_4194304a_wireless_laser_target_hatch", "4,194,304安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_4a_wireless_energy_input_hatch", "4安§5EV§r无线能源仓")
-        addCN("block.gtmoremachine.ev_4a_wireless_energy_output_hatch", "4安§5EV§r无线动力仓")
-        addCN("block.gtmoremachine.ev_64a_wireless_energy_input_hatch", "64安§5EV§r无线能源仓")
-        addCN("block.gtmoremachine.ev_64a_wireless_energy_output_hatch", "64安§5EV§r无线动力仓")
-        addCN("block.gtmoremachine.ev_65536a_wireless_laser_source_hatch", "65536安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_65536a_wireless_laser_target_hatch", "65536安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_67108863a_wireless_laser_source_hatch", "67,108,863安§5EV§r无线激光源仓")
-        addCN("block.gtmoremachine.ev_67108863a_wireless_laser_target_hatch", "67,108,863安§5EV§r无线激光靶仓")
-        addCN("block.gtmoremachine.ev_capacity_component", "无线电网容量组件 (§5EV§r)")
-        addCN("block.gtmoremachine.ev_energy_communication_unit", "能源通讯单元 (§5EV§r)")
-        addCN("block.gtmoremachine.ev_huge_dual_hatch", "§5EV§r巨型输入总成")
-        addCN("block.gtmoremachine.ev_huge_item_export_bus", "§5EV§r巨型输出总线")
-        addCN("block.gtmoremachine.ev_huge_item_import_bus", "§5EV§r巨型输入总线")
-        addCN("block.gtmoremachine.hv_16a_wireless_energy_input_hatch", "16安§6HV§r无线能源仓")
-        addCN("block.gtmoremachine.hv_16a_wireless_energy_output_hatch", "16安§6HV§r无线动力仓")
-        addCN("block.gtmoremachine.hv_2a_wireless_energy_input_hatch", "§6HV§r无线能源仓")
-        addCN("block.gtmoremachine.hv_2a_wireless_energy_output_hatch", "§6HV§r无线动力仓")
-        addCN("block.gtmoremachine.hv_4a_wireless_energy_input_hatch", "4安§6HV§r无线能源仓")
-        addCN("block.gtmoremachine.hv_4a_wireless_energy_output_hatch", "4安§6HV§r无线动力仓")
-        addCN("block.gtmoremachine.hv_capacity_component", "无线电网容量组件 (§6HV§r)")
-        addCN("block.gtmoremachine.hv_energy_communication_unit", "能源通讯单元 (§6HV§r)")
-        addCN("block.gtmoremachine.hv_huge_dual_hatch", "§6HV§r巨型输入总成")
-        addCN("block.gtmoremachine.hv_huge_item_export_bus", "§6HV§r巨型输出总线")
-        addCN("block.gtmoremachine.hv_huge_item_import_bus", "§6HV§r巨型输入总线")
-        addCN("block.gtmoremachine.iv_1024a_wireless_laser_source_hatch", "1024安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_1024a_wireless_laser_target_hatch", "1024安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_1048576a_wireless_laser_source_hatch", "1,048,576安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_1048576a_wireless_laser_target_hatch", "1,048,576安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_16384a_wireless_laser_source_hatch", "16384安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_16384a_wireless_laser_target_hatch", "16384安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_16777216a_wireless_laser_source_hatch", "16,777,216安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_16777216a_wireless_laser_target_hatch", "16,777,216安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_16a_wireless_energy_input_hatch", "16安§9IV§r无线能源仓")
-        addCN("block.gtmoremachine.iv_16a_wireless_energy_output_hatch", "16安§9IV§r无线动力仓")
-        addCN("block.gtmoremachine.iv_256a_wireless_laser_source_hatch", "256安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_256a_wireless_laser_target_hatch", "256安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_262144a_wireless_laser_source_hatch", "262,144安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_262144a_wireless_laser_target_hatch", "262,144安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_2a_wireless_energy_input_hatch", "§9IV§r无线能源仓")
-        addCN("block.gtmoremachine.iv_2a_wireless_energy_output_hatch", "§9IV§r无线动力仓")
-        addCN("block.gtmoremachine.iv_33554432a_wireless_laser_source_hatch", "33,554,432安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_33554432a_wireless_laser_target_hatch", "33,554,432安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_4096a_wireless_laser_source_hatch", "4096安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_4096a_wireless_laser_target_hatch", "4096安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_4194304a_wireless_laser_source_hatch", "4,194,304安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_4194304a_wireless_laser_target_hatch", "4,194,304安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_4a_wireless_energy_input_hatch", "4安§9IV§r无线能源仓")
-        addCN("block.gtmoremachine.iv_4a_wireless_energy_output_hatch", "4安§9IV§r无线动力仓")
-        addCN("block.gtmoremachine.iv_64a_wireless_energy_input_hatch", "64安§9IV§r无线能源仓")
-        addCN("block.gtmoremachine.iv_64a_wireless_energy_output_hatch", "64安§9IV§r无线动力仓")
-        addCN("block.gtmoremachine.iv_65536a_wireless_laser_source_hatch", "65536安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_65536a_wireless_laser_target_hatch", "65536安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_67108863a_wireless_laser_source_hatch", "67,108,863安§9IV§r无线激光源仓")
-        addCN("block.gtmoremachine.iv_67108863a_wireless_laser_target_hatch", "67,108,863安§9IV§r无线激光靶仓")
-        addCN("block.gtmoremachine.iv_capacity_component", "无线电网容量组件 (§9IV§r)")
-        addCN("block.gtmoremachine.iv_energy_communication_unit", "能源通讯单元 (§9IV§r)")
-        addCN("block.gtmoremachine.iv_huge_dual_hatch", "§9IV§r巨型输入总成")
-        addCN("block.gtmoremachine.iv_huge_item_export_bus", "§9IV§r巨型输出总线")
-        addCN("block.gtmoremachine.iv_huge_item_import_bus", "§9IV§r巨型输入总线")
-        addCN("block.gtmoremachine.luv_1024a_wireless_laser_source_hatch", "1024安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_1024a_wireless_laser_target_hatch", "1024安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_1048576a_wireless_laser_source_hatch", "1,048,576安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_1048576a_wireless_laser_target_hatch", "1,048,576安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_16384a_wireless_laser_source_hatch", "16384安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_16384a_wireless_laser_target_hatch", "16384安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_16777216a_wireless_laser_source_hatch", "16,777,216安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_16777216a_wireless_laser_target_hatch", "16,777,216安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_16a_wireless_energy_input_hatch", "16安§dLuV§r无线能源仓")
-        addCN("block.gtmoremachine.luv_16a_wireless_energy_output_hatch", "16安§dLuV§r无线动力仓")
-        addCN("block.gtmoremachine.luv_256a_wireless_laser_source_hatch", "256安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_256a_wireless_laser_target_hatch", "256安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_262144a_wireless_laser_source_hatch", "262,144安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_262144a_wireless_laser_target_hatch", "262,144安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_2a_wireless_energy_input_hatch", "§dLuV§r无线能源仓")
-        addCN("block.gtmoremachine.luv_2a_wireless_energy_output_hatch", "§dLuV§r无线动力仓")
-        addCN("block.gtmoremachine.luv_33554432a_wireless_laser_source_hatch", "33,554,432安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_33554432a_wireless_laser_target_hatch", "33,554,432安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_4096a_wireless_laser_source_hatch", "4096安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_4096a_wireless_laser_target_hatch", "4096安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_4194304a_wireless_laser_source_hatch", "4,194,304安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_4194304a_wireless_laser_target_hatch", "4,194,304安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_4a_wireless_energy_input_hatch", "4安§dLuV§r无线能源仓")
-        addCN("block.gtmoremachine.luv_4a_wireless_energy_output_hatch", "4安§dLuV§r无线动力仓")
-        addCN("block.gtmoremachine.luv_64a_wireless_energy_input_hatch", "64安§dLuV§r无线能源仓")
-        addCN("block.gtmoremachine.luv_64a_wireless_energy_output_hatch", "64安§dLuV§r无线动力仓")
-        addCN("block.gtmoremachine.luv_65536a_wireless_laser_source_hatch", "65536安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_65536a_wireless_laser_target_hatch", "65536安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_67108863a_wireless_laser_source_hatch", "67,108,863安§dLuV§r无线激光源仓")
-        addCN("block.gtmoremachine.luv_67108863a_wireless_laser_target_hatch", "67,108,863安§dLuV§r无线激光靶仓")
-        addCN("block.gtmoremachine.luv_capacity_component", "无线电网容量组件 (§dLuV§r)")
-        addCN("block.gtmoremachine.luv_energy_communication_unit", "能源通讯单元 (§dLuV§r)")
-        addCN("block.gtmoremachine.luv_huge_dual_hatch", "§dLuV§r巨型输入总成")
-        addCN("block.gtmoremachine.luv_huge_item_export_bus", "§dLuV§r巨型输出总线")
-        addCN("block.gtmoremachine.luv_huge_item_import_bus", "§dLuV§r巨型输入总线")
-        addCN("block.gtmoremachine.luv_programmablec_dualhatch", "§dLuV§r巨型可编程仓")
-        addCN("block.gtmoremachine.luv_programmablec_hatch", "§dLuV§r可编程仓")
-        addCN("block.gtmoremachine.lv_16a_wireless_energy_input_hatch", "16安§7LV§r无线能源仓")
-        addCN("block.gtmoremachine.lv_16a_wireless_energy_output_hatch", "16安§7LV§r无线动力仓")
-        addCN("block.gtmoremachine.lv_2a_wireless_energy_input_hatch", "§7LV§r无线能源仓")
-        addCN("block.gtmoremachine.lv_2a_wireless_energy_output_hatch", "§7LV§r无线动力仓")
-        addCN("block.gtmoremachine.lv_4a_wireless_energy_input_hatch", "4安§7LV§r无线能源仓")
-        addCN("block.gtmoremachine.lv_4a_wireless_energy_output_hatch", "4安§7LV§r无线动力仓")
-        addCN("block.gtmoremachine.lv_capacity_component", "无线电网容量组件 (§7LV§r)")
-        addCN("block.gtmoremachine.lv_energy_communication_unit", " 能源通讯单元 (§7LV§r)")
-        addCN("block.gtmoremachine.lv_huge_dual_hatch", "§7LV§r巨型输入总成")
-        addCN("block.gtmoremachine.lv_huge_item_export_bus", "§7LV§r巨型输出总线")
-        addCN("block.gtmoremachine.lv_huge_item_import_bus", "§7LV§r巨型输入总线")
-        addCN("block.gtmoremachine.max_1024a_wireless_laser_source_hatch", "1024安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_1024a_wireless_laser_target_hatch", "1024安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_1048576a_wireless_laser_source_hatch", "1,048,576安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_1048576a_wireless_laser_target_hatch", "1,048,576安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_16384a_wireless_laser_source_hatch", "16384安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_16384a_wireless_laser_target_hatch", "16384安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_16777216a_wireless_laser_source_hatch", "16,777,216安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_16777216a_wireless_laser_target_hatch", "16,777,216安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_16a_wireless_energy_input_hatch", "16安§c§lMAX§r无线能源仓")
-        addCN("block.gtmoremachine.max_16a_wireless_energy_output_hatch", "16安§c§lMAX§r无线动力仓")
-        addCN("block.gtmoremachine.max_256a_wireless_laser_source_hatch", "256安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_256a_wireless_laser_target_hatch", "256安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_262144a_wireless_laser_source_hatch", "262,144安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_262144a_wireless_laser_target_hatch", "262,144安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_2a_wireless_energy_input_hatch", "§c§lMAX§r无线能源仓")
-        addCN("block.gtmoremachine.max_2a_wireless_energy_output_hatch", "§c§lMAX§r无线动力仓")
-        addCN("block.gtmoremachine.max_33554432a_wireless_laser_source_hatch", "33,554,432安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_33554432a_wireless_laser_target_hatch", "33,554,432安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_4096a_wireless_laser_source_hatch", "4096安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_4096a_wireless_laser_target_hatch", "4096安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_4194304a_wireless_laser_source_hatch", "4,194,304安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_4194304a_wireless_laser_target_hatch", "4,194,304安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_4a_wireless_energy_input_hatch", "4安§c§lMAX§r无线能源仓")
-        addCN("block.gtmoremachine.max_4a_wireless_energy_output_hatch", "4安§c§lMAX§r无线动力仓")
-        addCN("block.gtmoremachine.max_64a_wireless_energy_input_hatch", "64安§c§lMAX§r无线能源仓")
-        addCN("block.gtmoremachine.max_64a_wireless_energy_output_hatch", "64安§c§lMAX§r无线动力仓")
-        addCN("block.gtmoremachine.max_65536a_wireless_laser_source_hatch", "65536安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_65536a_wireless_laser_target_hatch", "65536安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_67108863a_wireless_laser_source_hatch", "67,108,863安§c§lMAX§r无线激光源仓")
-        addCN("block.gtmoremachine.max_67108863a_wireless_laser_target_hatch", "67,108,863安§c§lMAX§r无线激光靶仓")
-        addCN("block.gtmoremachine.max_capacity_component", "无线电网容量组件 (§c§lMAX§r)")
-        addCN("block.gtmoremachine.max_energy_communication_unit", "能源通讯单元 (§c§lMAX§r)")
-        addCN("block.gtmoremachine.max_huge_dual_hatch", "§c§lMAX§r巨型输入总成")
-        addCN("block.gtmoremachine.max_huge_item_export_bus", "§c§lMAX§r巨型输出总线")
-        addCN("block.gtmoremachine.max_huge_item_import_bus", "§c§lMAX§r巨型输入总线")
-        addCN("block.gtmoremachine.max_programmablec_dualhatch", "§c§lMAX§r巨型可编程仓")
-        addCN("block.gtmoremachine.max_programmablec_hatch", "§c§lMAX§r可编程仓")
-        addCN("block.gtmoremachine.me_export_buffer", "ME输出总成")
-        addCN("block.gtmoremachine.mv_16a_wireless_energy_input_hatch", "16安§bMV§r无线能源仓")
-        addCN("block.gtmoremachine.mv_16a_wireless_energy_output_hatch", "16安§bMV§r无线动力仓")
-        addCN("block.gtmoremachine.mv_2a_wireless_energy_input_hatch", "§bMV§r无线能源仓")
-        addCN("block.gtmoremachine.mv_2a_wireless_energy_output_hatch", "§bMV§r无线动力仓")
-        addCN("block.gtmoremachine.mv_4a_wireless_energy_input_hatch", "4安§bMV§r无线能源仓")
-        addCN("block.gtmoremachine.mv_4a_wireless_energy_output_hatch", "4安§bMV§r无线动力仓")
-        addCN("block.gtmoremachine.mv_capacity_component", "无线电网容量组件 (§bMV§r)")
-        addCN("block.gtmoremachine.mv_energy_communication_unit", "能源通讯单元 (§bMV§r)")
-        addCN("block.gtmoremachine.mv_huge_dual_hatch", "§bMV§r巨型输入总成")
-        addCN("block.gtmoremachine.mv_huge_item_export_bus", "§bMV§r巨型输出总线")
-        addCN("block.gtmoremachine.mv_huge_item_import_bus", "§bMV§r巨型输入总线")
-        addCN("block.gtmoremachine.opv_1024a_wireless_laser_source_hatch", "1024安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_1024a_wireless_laser_target_hatch", "1024安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_1048576a_wireless_laser_source_hatch", "1,048,576安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_1048576a_wireless_laser_target_hatch", "1,048,576安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_16384a_wireless_laser_source_hatch", "16384安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_16384a_wireless_laser_target_hatch", "16384安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_16777216a_wireless_laser_source_hatch", "16,777,216安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_16777216a_wireless_laser_target_hatch", "16,777,216安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_16a_wireless_energy_input_hatch", "16安§9§lOpV§r无线能源仓")
-        addCN("block.gtmoremachine.opv_16a_wireless_energy_output_hatch", "16安§9§lOpV§r无线动力仓")
-        addCN("block.gtmoremachine.opv_256a_wireless_laser_source_hatch", "256安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_256a_wireless_laser_target_hatch", "256安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_262144a_wireless_laser_source_hatch", "262,144安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_262144a_wireless_laser_target_hatch", "262,144安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_2a_wireless_energy_input_hatch", "§9§lOpV§r无线能源仓")
-        addCN("block.gtmoremachine.opv_2a_wireless_energy_output_hatch", "§9§lOpV§r无线动力仓")
-        addCN("block.gtmoremachine.opv_33554432a_wireless_laser_source_hatch", "33,554,432安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_33554432a_wireless_laser_target_hatch", "33,554,432安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_4096a_wireless_laser_source_hatch", "4096安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_4096a_wireless_laser_target_hatch", "4096安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_4194304a_wireless_laser_source_hatch", "4,194,304安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_4194304a_wireless_laser_target_hatch", "4,194,304安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_4a_wireless_energy_input_hatch", "4安§9§lOpV§r无线能源仓")
-        addCN("block.gtmoremachine.opv_4a_wireless_energy_output_hatch", "4安§9§lOpV§r无线动力仓")
-        addCN("block.gtmoremachine.opv_64a_wireless_energy_input_hatch", "64安§9§lOpV§r无线能源仓")
-        addCN("block.gtmoremachine.opv_64a_wireless_energy_output_hatch", "64安§9§lOpV§r无线动力仓")
-        addCN("block.gtmoremachine.opv_65536a_wireless_laser_source_hatch", "65536安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_65536a_wireless_laser_target_hatch", "65536安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_67108863a_wireless_laser_source_hatch", "67,108,863安§9§lOpV§r无线激光源仓")
-        addCN("block.gtmoremachine.opv_67108863a_wireless_laser_target_hatch", "67,108,863安§9§lOpV§r无线激光靶仓")
-        addCN("block.gtmoremachine.opv_capacity_component", "无线电网容量组件 (§9§lOpV§r)")
-        addCN("block.gtmoremachine.opv_energy_communication_unit", "能源通讯单元 (§9§lOpV§r)")
-        addCN("block.gtmoremachine.opv_huge_dual_hatch", "§9§lOpV§r巨型输入总成")
-        addCN("block.gtmoremachine.opv_huge_item_export_bus", "§9§lOpV§r巨型输出总线")
-        addCN("block.gtmoremachine.opv_huge_item_import_bus", "§9§lOpV§r巨型输入总线")
-        addCN("block.gtmoremachine.opv_programmablec_dualhatch", "§9§lOpV§r巨型可编程仓")
-        addCN("block.gtmoremachine.opv_programmablec_hatch", "§9§lOpV§r可编程仓")
-        addCN("block.gtmoremachine.uev_1024a_wireless_laser_source_hatch", "1024安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_1024a_wireless_laser_target_hatch", "1024安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_1048576a_wireless_laser_source_hatch", "1,048,576安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_1048576a_wireless_laser_target_hatch", "1,048,576安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_16384a_wireless_laser_source_hatch", "16384安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_16384a_wireless_laser_target_hatch", "16384安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_16777216a_wireless_laser_source_hatch", "16,777,216安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_16777216a_wireless_laser_target_hatch", "16,777,216安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_16a_wireless_energy_input_hatch", "16安§aUEV§r无线能源仓")
-        addCN("block.gtmoremachine.uev_16a_wireless_energy_output_hatch", "16安§aUEV§r无线动力仓")
-        addCN("block.gtmoremachine.uev_256a_wireless_laser_source_hatch", "256安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_256a_wireless_laser_target_hatch", "256安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_262144a_wireless_laser_source_hatch", "262,144安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_262144a_wireless_laser_target_hatch", "262,144安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_2a_wireless_energy_input_hatch", "§aUEV§r无线能源仓")
-        addCN("block.gtmoremachine.uev_2a_wireless_energy_output_hatch", "§aUEV§r无线动力仓")
-        addCN("block.gtmoremachine.uev_33554432a_wireless_laser_source_hatch", "33,554,432安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_33554432a_wireless_laser_target_hatch", "33,554,432安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_4096a_wireless_laser_source_hatch", "4096安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_4096a_wireless_laser_target_hatch", "4096安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_4194304a_wireless_laser_source_hatch", "4,194,304安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_4194304a_wireless_laser_target_hatch", "4,194,304安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_4a_wireless_energy_input_hatch", "4安§aUEV§r无线能源仓")
-        addCN("block.gtmoremachine.uev_4a_wireless_energy_output_hatch", "4安§aUEV§r无线动力仓")
-        addCN("block.gtmoremachine.uev_64a_wireless_energy_input_hatch", "64安§aUEV§r无线能源仓")
-        addCN("block.gtmoremachine.uev_64a_wireless_energy_output_hatch", "64安§aUEV§r无线动力仓")
-        addCN("block.gtmoremachine.uev_65536a_wireless_laser_source_hatch", "65536安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_65536a_wireless_laser_target_hatch", "65536安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_67108863a_wireless_laser_source_hatch", "67,108,863安§aUEV§r无线激光源仓")
-        addCN("block.gtmoremachine.uev_67108863a_wireless_laser_target_hatch", "67,108,863安§aUEV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uev_capacity_component", "无线电网容量组件 (§aUEV§r)")
-        addCN("block.gtmoremachine.uev_energy_communication_unit", "能源通讯单元 (§aUEV§r)")
-        addCN("block.gtmoremachine.uev_huge_dual_hatch", "§aUEV§r巨型输入总成")
-        addCN("block.gtmoremachine.uev_huge_item_export_bus", "§aUEV§r巨型输出总线")
-        addCN("block.gtmoremachine.uev_huge_item_import_bus", "§aUEV§r巨型输入总线")
-        addCN("block.gtmoremachine.uev_programmablec_dualhatch", "§aUEV§r巨型可编程仓")
-        addCN("block.gtmoremachine.uev_programmablec_hatch", "§aUEV§r可编程仓")
-        addCN("block.gtmoremachine.uhv_1024a_wireless_laser_source_hatch", "1024安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_1024a_wireless_laser_target_hatch", "1024安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_1048576a_wireless_laser_source_hatch", "1,048,576安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_1048576a_wireless_laser_target_hatch", "1,048,576安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_16384a_wireless_laser_source_hatch", "16384安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_16384a_wireless_laser_target_hatch", "16384安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_16777216a_wireless_laser_source_hatch", "16,777,216安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_16777216a_wireless_laser_target_hatch", "16,777,216安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_16a_wireless_energy_input_hatch", "16安§4UHV§r无线能源仓")
-        addCN("block.gtmoremachine.uhv_16a_wireless_energy_output_hatch", "16安§4UHV§r无线动力仓")
-        addCN("block.gtmoremachine.uhv_256a_wireless_laser_source_hatch", "256安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_256a_wireless_laser_target_hatch", "256安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_262144a_wireless_laser_source_hatch", "262,144安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_262144a_wireless_laser_target_hatch", "262,144安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_2a_wireless_energy_input_hatch", "§4UHV§r无线能源仓")
-        addCN("block.gtmoremachine.uhv_2a_wireless_energy_output_hatch", "§4UHV§r无线动力仓")
-        addCN("block.gtmoremachine.uhv_33554432a_wireless_laser_source_hatch", "33,554,432安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_33554432a_wireless_laser_target_hatch", "33,554,432安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_4096a_wireless_laser_source_hatch", "4096安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_4096a_wireless_laser_target_hatch", "4096安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_4194304a_wireless_laser_source_hatch", "4,194,304安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_4194304a_wireless_laser_target_hatch", "4,194,304安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_4a_wireless_energy_input_hatch", "4安§4UHV§r无线能源仓")
-        addCN("block.gtmoremachine.uhv_4a_wireless_energy_output_hatch", "4安§4UHV§r无线动力仓")
-        addCN("block.gtmoremachine.uhv_64a_wireless_energy_input_hatch", "64安§4UHV§r无线能源仓")
-        addCN("block.gtmoremachine.uhv_64a_wireless_energy_output_hatch", "64安§4UHV§r无线动力仓")
-        addCN("block.gtmoremachine.uhv_65536a_wireless_laser_source_hatch", "65536安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_65536a_wireless_laser_target_hatch", "65536安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_67108863a_wireless_laser_source_hatch", "67,108,863安§4UHV§r无线激光源仓")
-        addCN("block.gtmoremachine.uhv_67108863a_wireless_laser_target_hatch", "67,108,863安§4UHV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uhv_capacity_component", "无线电网容量组件 (§4UHV§r)")
-        addCN("block.gtmoremachine.uhv_energy_communication_unit", "能源通讯单元 (§4UHV§r)")
-        addCN("block.gtmoremachine.uhv_huge_dual_hatch", "§4UHV§r巨型输入总成")
-        addCN("block.gtmoremachine.uhv_huge_item_export_bus", "§4UHV§r巨型输出总线")
-        addCN("block.gtmoremachine.uhv_huge_item_import_bus", "§4UHV§r巨型输入总线")
-        addCN("block.gtmoremachine.uhv_programmablec_dualhatch", "§4UHV§r巨型可编程仓")
-        addCN("block.gtmoremachine.uhv_programmablec_hatch", "§4UHV§r可编程仓")
-        addCN("block.gtmoremachine.uiv_1024a_wireless_laser_source_hatch", "1024安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_1024a_wireless_laser_target_hatch", "1024安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_1048576a_wireless_laser_source_hatch", "1,048,576安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_1048576a_wireless_laser_target_hatch", "1,048,576安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_16384a_wireless_laser_source_hatch", "16384安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_16384a_wireless_laser_target_hatch", "16384安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_16777216a_wireless_laser_source_hatch", "16,777,216安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_16777216a_wireless_laser_target_hatch", "16,777,216安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_16a_wireless_energy_input_hatch", "16安§2UIV§r无线能源仓")
-        addCN("block.gtmoremachine.uiv_16a_wireless_energy_output_hatch", "16安§2UIV§r无线动力仓")
-        addCN("block.gtmoremachine.uiv_256a_wireless_laser_source_hatch", "256安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_256a_wireless_laser_target_hatch", "256安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_262144a_wireless_laser_source_hatch", "262,144安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_262144a_wireless_laser_target_hatch", "262,144安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_2a_wireless_energy_input_hatch", "§2UIV§r无线能源仓")
-        addCN("block.gtmoremachine.uiv_2a_wireless_energy_output_hatch", "§2UIV§r无线动力仓")
-        addCN("block.gtmoremachine.uiv_33554432a_wireless_laser_source_hatch", "33,554,432安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_33554432a_wireless_laser_target_hatch", "33,554,432安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_4096a_wireless_laser_source_hatch", "4096安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_4096a_wireless_laser_target_hatch", "4096安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_4194304a_wireless_laser_source_hatch", "4,194,304安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_4194304a_wireless_laser_target_hatch", "4,194,304安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_4a_wireless_energy_input_hatch", "4安§2UIV§r无线能源仓")
-        addCN("block.gtmoremachine.uiv_4a_wireless_energy_output_hatch", "4安§2UIV§r无线动力仓")
-        addCN("block.gtmoremachine.uiv_64a_wireless_energy_input_hatch", "64安§2UIV§r无线能源仓")
-        addCN("block.gtmoremachine.uiv_64a_wireless_energy_output_hatch", "64安§2UIV§r无线动力仓")
-        addCN("block.gtmoremachine.uiv_65536a_wireless_laser_source_hatch", "65536安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_65536a_wireless_laser_target_hatch", "65536安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_67108863a_wireless_laser_source_hatch", "67,108,863安§2UIV§r无线激光源仓")
-        addCN("block.gtmoremachine.uiv_67108863a_wireless_laser_target_hatch", "67,108,863安§2UIV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uiv_capacity_component", "无线电网容量组件 (§2UIV§r)")
-        addCN("block.gtmoremachine.uiv_energy_communication_unit", "能源通讯单元 (§2UIV§r)")
-        addCN("block.gtmoremachine.uiv_huge_dual_hatch", "§2UIV§r巨型输入总成")
-        addCN("block.gtmoremachine.uiv_huge_item_export_bus", "§2UIV§r巨型输出总线")
-        addCN("block.gtmoremachine.uiv_huge_item_import_bus", "§2UIV§r巨型输入总线")
-        addCN("block.gtmoremachine.uiv_programmablec_dualhatch", "§2UIV§r巨型可编程仓")
-        addCN("block.gtmoremachine.uiv_programmablec_hatch", "§2UIV§r可编程仓")
-        addCN("block.gtmoremachine.ulv_huge_dual_hatch", "§8ULV§r巨型输入总成")
-        addCN("block.gtmoremachine.ulv_huge_item_export_bus", "§8ULV§r巨型输出总线")
-        addCN("block.gtmoremachine.ulv_huge_item_import_bus", "§8ULV§r巨型输入总线")
-        addCN("block.gtmoremachine.uv_1024a_wireless_laser_source_hatch", "1024安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_1024a_wireless_laser_target_hatch", "1024安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_1048576a_wireless_laser_source_hatch", "1,048,576安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_1048576a_wireless_laser_target_hatch", "1,048,576安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_16384a_wireless_laser_source_hatch", "16384安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_16384a_wireless_laser_target_hatch", "16384安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_16777216a_wireless_laser_source_hatch", "16,777,216安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_16777216a_wireless_laser_target_hatch", "16,777,216安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_16a_wireless_energy_input_hatch", "16安§3UV§r无线能源仓")
-        addCN("block.gtmoremachine.uv_16a_wireless_energy_output_hatch", "16安§3UV§r无线动力仓")
-        addCN("block.gtmoremachine.uv_256a_wireless_laser_source_hatch", "256安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_256a_wireless_laser_target_hatch", "256安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_262144a_wireless_laser_source_hatch", "262,144安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_262144a_wireless_laser_target_hatch", "262,144安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_2a_wireless_energy_input_hatch", "§3UV§r无线能源仓")
-        addCN("block.gtmoremachine.uv_2a_wireless_energy_output_hatch", "§3UV§r无线动力仓")
-        addCN("block.gtmoremachine.uv_33554432a_wireless_laser_source_hatch", "33,554,432安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_33554432a_wireless_laser_target_hatch", "33,554,432安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_4096a_wireless_laser_source_hatch", "4096安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_4096a_wireless_laser_target_hatch", "4096安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_4194304a_wireless_laser_source_hatch", "4,194,304安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_4194304a_wireless_laser_target_hatch", "4,194,304安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_4a_wireless_energy_input_hatch", "4安§3UV§r无线能源仓")
-        addCN("block.gtmoremachine.uv_4a_wireless_energy_output_hatch", "4安§3UV§r无线动力仓")
-        addCN("block.gtmoremachine.uv_64a_wireless_energy_input_hatch", "64安§3UV§r无线能源仓")
-        addCN("block.gtmoremachine.uv_64a_wireless_energy_output_hatch", "64安§3UV§r无线动力仓")
-        addCN("block.gtmoremachine.uv_65536a_wireless_laser_source_hatch", "65536安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_65536a_wireless_laser_target_hatch", "65536安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_67108863a_wireless_laser_source_hatch", "67,108,863安§3UV§r无线激光源仓")
-        addCN("block.gtmoremachine.uv_67108863a_wireless_laser_target_hatch", "67,108,863安§3UV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uv_capacity_component", "无线电网容量组件 (§3UV§r)")
-        addCN("block.gtmoremachine.uv_energy_communication_unit", "能源通讯单元 (§3UV§r)")
-        addCN("block.gtmoremachine.uv_huge_dual_hatch", "§3UV§r巨型输入总成")
-        addCN("block.gtmoremachine.uv_huge_item_export_bus", "§3UV§r巨型输出总线")
-        addCN("block.gtmoremachine.uv_huge_item_import_bus", "§3UV§r巨型输入总线")
-        addCN("block.gtmoremachine.uv_programmablec_dualhatch", "§3UV§r巨型可编程仓")
-        addCN("block.gtmoremachine.uv_programmablec_hatch", "§3UV§r可编程仓")
-        addCN("block.gtmoremachine.uxv_1024a_wireless_laser_source_hatch", "1024安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_1024a_wireless_laser_target_hatch", "1024安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_1048576a_wireless_laser_source_hatch", "1,048,576安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_1048576a_wireless_laser_target_hatch", "1,048,576安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_16384a_wireless_laser_source_hatch", "16384安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_16384a_wireless_laser_target_hatch", "16384安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_16777216a_wireless_laser_source_hatch", "16,777,216安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_16777216a_wireless_laser_target_hatch", "16,777,216安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_16a_wireless_energy_input_hatch", "16安§eUXV§r无线能源仓")
-        addCN("block.gtmoremachine.uxv_16a_wireless_energy_output_hatch", "16安§eUXV§r无线动力仓")
-        addCN("block.gtmoremachine.uxv_256a_wireless_laser_source_hatch", "256安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_256a_wireless_laser_target_hatch", "256安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_262144a_wireless_laser_source_hatch", "262,144安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_262144a_wireless_laser_target_hatch", "262,144安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_2a_wireless_energy_input_hatch", "§eUXV§r无线能源仓")
-        addCN("block.gtmoremachine.uxv_2a_wireless_energy_output_hatch", "§eUXV§r无线动力仓")
-        addCN("block.gtmoremachine.uxv_33554432a_wireless_laser_source_hatch", "33,554,432安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_33554432a_wireless_laser_target_hatch", "33,554,432安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_4096a_wireless_laser_source_hatch", "4096安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_4096a_wireless_laser_target_hatch", "4096安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_4194304a_wireless_laser_source_hatch", "4,194,304安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_4194304a_wireless_laser_target_hatch", "4,194,304安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_4a_wireless_energy_input_hatch", "4安§eUXV§r无线能源仓")
-        addCN("block.gtmoremachine.uxv_4a_wireless_energy_output_hatch", "4安§eUXV§r无线动力仓")
-        addCN("block.gtmoremachine.uxv_64a_wireless_energy_input_hatch", "64安§eUXV§r无线能源仓")
-        addCN("block.gtmoremachine.uxv_64a_wireless_energy_output_hatch", "64安§eUXV§r无线动力仓")
-        addCN("block.gtmoremachine.uxv_65536a_wireless_laser_source_hatch", "65536安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_65536a_wireless_laser_target_hatch", "65536安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_67108863a_wireless_laser_source_hatch", "67,108,863安§eUXV§r无线激光源仓")
-        addCN("block.gtmoremachine.uxv_67108863a_wireless_laser_target_hatch", "67,108,863安§eUXV§r无线激光靶仓")
-        addCN("block.gtmoremachine.uxv_capacity_component", "无线电网容量组件 (§eUXV§r)")
-        addCN("block.gtmoremachine.uxv_energy_communication_unit", "能源通讯单元 (§eUXV§r)")
-        addCN("block.gtmoremachine.uxv_huge_dual_hatch", "§eUXV§r巨型输入总成")
-        addCN("block.gtmoremachine.uxv_huge_item_export_bus", "§eUXV§r巨型输出总线")
-        addCN("block.gtmoremachine.uxv_huge_item_import_bus", "§eUXV§r巨型输入总线")
-        addCN("block.gtmoremachine.uxv_programmablec_dualhatch", "§eUXV§r巨型可编程仓")
-        addCN("block.gtmoremachine.uxv_programmablec_hatch", "§eUXV§r可编程仓")
-        addCN("block.gtmoremachine.wireless_computation_receiver_hatch", "无线算力数据靶仓")
-        addCN("block.gtmoremachine.wireless_computation_transmitter_hatch", "无线算力数据源仓")
-        addCN("block.gtmoremachine.wireless_cwu_monitor", "无线算力监视器")
-        addCN("block.gtmoremachine.wireless_cwu_receiver_hatch", "算力数据网络下行仓")
-        addCN("block.gtmoremachine.wireless_cwu_transmitter_hatch", "算力数据网络上行仓")
-        addCN("block.gtmoremachine.wireless_energy_interface", "无线能源接口")
-        addCN("block.gtmoremachine.wireless_energy_monitor", "无线能源监视器")
-        addCN("block.gtmoremachine.wirelsess_energy_demodulation_hub", "无线电网解调枢纽")
-        addCN("block.gtmoremachine.wirelsess_energy_dimensional_relay_node", "无线电网维度中继节点")
-        addCN("block.gtmoremachine.zpm_1024a_wireless_laser_source_hatch", "1024安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_1024a_wireless_laser_target_hatch", "1024安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_1048576a_wireless_laser_source_hatch", "1,048,576安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_1048576a_wireless_laser_target_hatch", "1,048,576安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_16384a_wireless_laser_source_hatch", "16384安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_16384a_wireless_laser_target_hatch", "16384安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_16777216a_wireless_laser_source_hatch", "16,777,216安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_16777216a_wireless_laser_target_hatch", "16,777,216安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_16a_wireless_energy_input_hatch", "16安§cZPM§r无线能源仓")
-        addCN("block.gtmoremachine.zpm_16a_wireless_energy_output_hatch", "16安§cZPM§r无线动力仓")
-        addCN("block.gtmoremachine.zpm_256a_wireless_laser_source_hatch", "256安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_256a_wireless_laser_target_hatch", "256安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_262144a_wireless_laser_source_hatch", "262,144安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_262144a_wireless_laser_target_hatch", "262,144安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_2a_wireless_energy_input_hatch", "§cZPM§r无线能源仓")
-        addCN("block.gtmoremachine.zpm_2a_wireless_energy_output_hatch", "§cZPM§r无线动力仓")
-        addCN("block.gtmoremachine.zpm_33554432a_wireless_laser_source_hatch", "33,554,432安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_33554432a_wireless_laser_target_hatch", "33,554,432安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_4096a_wireless_laser_source_hatch", "4096安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_4096a_wireless_laser_target_hatch", "4096安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_4194304a_wireless_laser_source_hatch", "4,194,304安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_4194304a_wireless_laser_target_hatch", "4,194,304安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_4a_wireless_energy_input_hatch", "4安§cZPM§r无线能源仓")
-        addCN("block.gtmoremachine.zpm_4a_wireless_energy_output_hatch", "4安§cZPM§r无线动力仓")
-        addCN("block.gtmoremachine.zpm_64a_wireless_energy_input_hatch", "64安§cZPM§r无线能源仓")
-        addCN("block.gtmoremachine.zpm_64a_wireless_energy_output_hatch", "64安§cZPM§r无线动力仓")
-        addCN("block.gtmoremachine.zpm_65536a_wireless_laser_source_hatch", "65536安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_65536a_wireless_laser_target_hatch", "65536安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_67108863a_wireless_laser_source_hatch", "67,108,863安§cZPM§r无线激光源仓")
-        addCN("block.gtmoremachine.zpm_67108863a_wireless_laser_target_hatch", "67,108,863安§cZPM§r无线激光靶仓")
-        addCN("block.gtmoremachine.zpm_capacity_component", "无线电网容量组件 (§cZPM§r)")
-        addCN("block.gtmoremachine.zpm_energy_communication_unit", "能源通讯单元 (§cZPM§r)")
-        addCN("block.gtmoremachine.zpm_huge_dual_hatch", "§cZPM§r巨型输入总成")
-        addCN("block.gtmoremachine.zpm_huge_item_export_bus", "§cZPM§r巨型输出总线")
-        addCN("block.gtmoremachine.zpm_huge_item_import_bus", "§cZPM§r巨型输入总线")
-        addCN("block.gtmoremachine.zpm_programmablec_dualhatch", "§cZPM§r巨型可编程仓")
-        addCN("block.gtmoremachine.zpm_programmablec_hatch", "§cZPM§r可编程仓")
-        addCN("item.gtmoremachine.advanced_terminal", "§b高级终端")
-        addCN("item.gtmoremachine.advanced_wireless_fluid_transfer_cover", "§b进阶无线流体传输覆盖板")
-        addCN("item.gtmoremachine.advanced_wireless_item_transfer_cover", "§b进阶无线物品传输覆盖板")
-        addCN("item.gtmoremachine.creative_energy_cover", "创造能源覆盖板")
-        addCN("item.gtmoremachine.creative_fluid_cell", "创造流体单元")
-        addCN("item.gtmoremachine.ev_4a_wireless_energy_receive_cover", "4安§5EV§r无线能源接收器")
-        addCN("item.gtmoremachine.ev_wireless_energy_receive_cover", "§5EV§r无线能源接收器")
-        addCN("item.gtmoremachine.hv_4a_wireless_energy_receive_cover", "4安§6HV§r无线能源接收器")
-        addCN("item.gtmoremachine.hv_wireless_energy_receive_cover", "§6HV§r无线能源接收器")
-        addCN("item.gtmoremachine.iv_4a_wireless_energy_receive_cover", "4安§9IV§r无线能源接收器")
-        addCN("item.gtmoremachine.iv_wireless_energy_receive_cover", "§9IV§r无线能源接收器")
-        addCN("item.gtmoremachine.luv_4a_wireless_energy_receive_cover", "4安§dLuV§r无线能源接收器")
-        addCN("item.gtmoremachine.luv_wireless_energy_receive_cover", "§dLuV§r无线能源接收器")
-        addCN("item.gtmoremachine.lv_4a_wireless_energy_receive_cover", "4安§7LV§r无线能源接收器")
-        addCN("item.gtmoremachine.lv_wireless_energy_receive_cover", "§7LV§r无线能源接收器")
-        addCN("item.gtmoremachine.mv_4a_wireless_energy_receive_cover", "4安§bMV§r无线能源接收器")
-        addCN("item.gtmoremachine.mv_wireless_energy_receive_cover", "§bMV§r无线能源接收器")
-        addCN("item.gtmoremachine.opv_4a_wireless_energy_receive_cover", "4安§9§lOpV§r无线能源接收器")
-        addCN("item.gtmoremachine.opv_wireless_energy_receive_cover", "§9§lOpV§r无线能源接收器")
-        addCN("item.gtmoremachine.programmable_cover", "可编程覆盖板")
-        addCN("item.gtmoremachine.uev_4a_wireless_energy_receive_cover", "4安§aUEV§r无线能源接收器")
-        addCN("item.gtmoremachine.uev_wireless_energy_receive_cover", "§aUEV§r无线能源接收器")
-        addCN("item.gtmoremachine.uhv_4a_wireless_energy_receive_cover", "4安§4UHV§r无线能源接收器")
-        addCN("item.gtmoremachine.uhv_wireless_energy_receive_cover", "§4UHV§r无线能源接收器")
-        addCN("item.gtmoremachine.uiv_4a_wireless_energy_receive_cover", "4安§2UIV§r无线能源接收器")
-        addCN("item.gtmoremachine.uiv_wireless_energy_receive_cover", "§2UIV§r无线能源接收器")
-        addCN("item.gtmoremachine.uv_4a_wireless_energy_receive_cover", "4安§3UV§r无线能源接收器")
-        addCN("item.gtmoremachine.uv_wireless_energy_receive_cover", "§3UV§r无线能源接收器")
-        addCN("item.gtmoremachine.uxv_4a_wireless_energy_receive_cover", "4安§eUXV§r无线能源接收器")
-        addCN("item.gtmoremachine.uxv_wireless_energy_receive_cover", "§eUXV§r无线能源接收器")
-        addCN("item.gtmoremachine.virtual_item_provider", "虚拟物品提供器")
-        addCN("item.gtmoremachine.virtual_item_provider_cell", "虚拟物品提供器元件")
-        addCN("item.gtmoremachine.wireless_energy_binding_tool", "无线能源绑定工具")
-        addCN("item.gtmoremachine.wireless_energy_terminal", "无线能源终端")
-        addCN("item.gtmoremachine.wireless_fluid_transfer_cover", "无线流体传输覆盖板")
-        addCN("item.gtmoremachine.wireless_item_transfer_cover", "无线物品传输覆盖板")
-        addCN("item.gtmoremachine.zpm_4a_wireless_energy_receive_cover", "4安§cZPM§r无线能源接收器")
-        addCN("item.gtmoremachine.zpm_wireless_energy_receive_cover", "§cZPM§r无线能源接收器")
-        addCN("itemGroup.gtmoremachine.creative", "创造物品")
-        addCN("itemGroup.gtmoremachine.more_machines", "更多机器")
-        addCN("itemGroup.gtmoremachine.wireless", "格雷无线科技")
-    }
-}
