+++ src/main/java/cn/qiuye/gtmoremachine/data/lang/WirelessItmeLang.kt
@@ -1,0 +1,59 @@
+package cn.qiuye.gtmoremachine.data.lang
+
+import cn.qiuye.gtmoremachine.data.lang.LangHandler.addCN
+
+// LV
+// MV
+// HV
+// EV
+// IV
+// LUV
+// ZPM
+// UV
+// UHV
+// UEV
+// UIV
+// UXV
+// OPV
+// MAX
+object WirelessItmeLang {
+
+    fun init() {
+        wirelessEnergyReceiveCover()
+        addCN("item.gtmoremachine.wireless_energy_binding_tool", "无线能源绑定工具")
+        addCN("item.gtmoremachine.wireless_energy_terminal", "无线能源终端")
+        addCN("item.gtmoremachine.wireless_fluid_transfer_cover", "无线流体传输覆盖板")
+        addCN("item.gtmoremachine.wireless_item_transfer_cover", "无线物品传输覆盖板")
+        addCN("item.gtmoremachine.advanced_wireless_fluid_transfer_cover", "§b进阶无线流体传输覆盖板")
+        addCN("item.gtmoremachine.advanced_wireless_item_transfer_cover", "§b进阶无线物品传输覆盖板")
+    }
+
+    fun wirelessEnergyReceiveCover() {
+        addCN("item.gtmoremachine.lv_4a_wireless_energy_receive_cover", "4安§7LV§r无线能源接收器")
+        addCN("item.gtmoremachine.lv_wireless_energy_receive_cover", "§7LV§r无线能源接收器")
+        addCN("item.gtmoremachine.mv_4a_wireless_energy_receive_cover", "4安§bMV§r无线能源接收器")
+        addCN("item.gtmoremachine.mv_wireless_energy_receive_cover", "§bMV§r无线能源接收器")
+        addCN("item.gtmoremachine.hv_4a_wireless_energy_receive_cover", "4安§6HV§r无线能源接收器")
+        addCN("item.gtmoremachine.hv_wireless_energy_receive_cover", "§6HV§r无线能源接收器")
+        addCN("item.gtmoremachine.ev_4a_wireless_energy_receive_cover", "4安§5EV§r无线能源接收器")
+        addCN("item.gtmoremachine.ev_wireless_energy_receive_cover", "§5EV§r无线能源接收器")
+        addCN("item.gtmoremachine.iv_4a_wireless_energy_receive_cover", "4安§9IV§r无线能源接收器")
+        addCN("item.gtmoremachine.iv_wireless_energy_receive_cover", "§9IV§r无线能源接收器")
+        addCN("item.gtmoremachine.luv_4a_wireless_energy_receive_cover", "4安§dLuV§r无线能源接收器")
+        addCN("item.gtmoremachine.luv_wireless_energy_receive_cover", "§dLuV§r无线能源接收器")
+        addCN("item.gtmoremachine.zpm_4a_wireless_energy_receive_cover", "4安§cZPM§r无线能源接收器")
+        addCN("item.gtmoremachine.zpm_wireless_energy_receive_cover", "§cZPM§r无线能源接收器")
+        addCN("item.gtmoremachine.uv_4a_wireless_energy_receive_cover", "4安§3UV§r无线能源接收器")
+        addCN("item.gtmoremachine.uv_wireless_energy_receive_cover", "§3UV§r无线能源接收器")
+        addCN("item.gtmoremachine.uhv_4a_wireless_energy_receive_cover", "4安§4UHV§r无线能源接收器")
+        addCN("item.gtmoremachine.uhv_wireless_energy_receive_cover", "§4UHV§r无线能源接收器")
+        addCN("item.gtmoremachine.uev_4a_wireless_energy_receive_cover", "4安§aUEV§r无线能源接收器")
+        addCN("item.gtmoremachine.uev_wireless_energy_receive_cover", "§aUEV§r无线能源接收器")
+        addCN("item.gtmoremachine.uiv_4a_wireless_energy_receive_cover", "4安§2UIV§r无线能源接收器")
+        addCN("item.gtmoremachine.uiv_wireless_energy_receive_cover", "§2UIV§r无线能源接收器")
+        addCN("item.gtmoremachine.uxv_4a_wireless_energy_receive_cover", "4安§eUXV§r无线能源接收器")
+        addCN("item.gtmoremachine.uxv_wireless_energy_receive_cover", "§eUXV§r无线能源接收器")
+        addCN("item.gtmoremachine.opv_4a_wireless_energy_receive_cover", "4安§9§lOpV§r无线能源接收器")
+        addCN("item.gtmoremachine.opv_wireless_energy_receive_cover", "§9§lOpV§r无线能源接收器")
+    }
+}
